Decouple CloudWatchReporter from ScheduledReporter.
This change Makes CloudWatchReporter a standalone Reporter
implementation. It has a report() method with the same signature as
ScheduledReporter#report, but it can be called at any time.

This is the non-backwards-compatible approach, as it changes
CloudWatchReporter to no longer be a ScheduledReporter. Since it's
already backwards-incompatible, I removed a lot of the deprecated fields
that were being kept for backwards-compatibility.

Also removed the builder as the only reason it was there was to support
new default values.